requirements.txt
pandas~=2.2.3
numpy~=2.2.4
matplotlib~=3.10.1
seaborn~=0.13.2
joblib~=1.5.0
scikit-learn
flask
gunicorn
Flask-Mail
python-dotenv
requests
python-dotenv
supabase
sqlalchemy
flask-sqlalchemy
flask-migrate
psycopg2-binary
openai
faiss-cpu
tiktoken
pyjwt
groq
markdown
bcrypt
plotly
